all: reimplement Store on top of KVDB instead of BoltDB
BoltDB is great but its massive use of mmap makes it quickly unusable on
32-bits machines like RaspberryPi. apec is using 3 stores:
- bleve
- store
- geocoder
I cannot touch the bleve one for now and the geocoder is size 128M while
the store one has reached the 512M mark. So migrating the store to
another backend might help the bleve store load.